Tests to illustrate issue 197...
Three tests are added. The first two may not be desirable specifications,
since passing them might cause a major version bump again.

The third test is an example of a possible scenario to mitigage confusion.
That would be my suggestion if passing the first two is too much of a
burden. Or something like that.