Warn user when `includesText` assertion should expect collapsable whitespace (#198)
* Tests to illustrate issue 197...

Three tests are added. The first two may not be desirable specifications,
since passing them might cause a major version bump again.

The third test is an example of a possible scenario to mitigage confusion.
That would be my suggestion if passing the first two is too much of a
burden. Or something like that.

* Notify the user if they pass extra whitespace into `.includesText()`...

This is my suggestion for addressing issue #197.

* Remove example tests

* Add a line breaks to user message

* Add warning about whitespace in API docs

* Update includes-text jest test

* Use the proper name of the class `RegExp` in includesText warning

* Change user message to appear in console as a warning instead

https://github.com/simplabs/qunit-dom/pull/198#issuecomment-440631014

* Add note to `hasText` documentation about whitespace collapsing.

https://github.com/simplabs/qunit-dom/pull/198#issuecomment-440631014

* Adjust test to ensure user is notified via console.warn