Expose a real module
This package defines an empty module for the convenience of typescript users, so they can say `import "qunit-dom"` to get the types loaded. It's not really needed for any actual code, because the code gets applied directly to QUnit's own objects.

But because this package declares itself an ember-addon, its own module namespace is expected to be defined by real modules in `treeForAddon` or `treeForAddonTestSupport`. Not a dynamic `define` in `vendor`.

`define` is most commonly used for things that aren't in an addons own module namespace, and that continues to work under embroider. But this more unusual case -- an addon `define`ing itself -- is not supported.

Without this change, attempting to `import "qunit-dom"` fails under embroider.