Do not default `target` to `document`.
The prior behavior was pretty broken, in the case where `assert.dom()`
is called (no arguments) `rootElement` will be `document` for non-Ember
users (because `rootElement` falls back to `document`), but our types do
not allow `document` for the `target` property of `DOMAssertions`.

It isn't _just_ a types issue though, things like
`findElement`/`findElements` actually error if `this.target` is not
`null`, a string, or `instanceof Element` (which `document` is not)!

This changes the defaulting to set `target` to the `rootElement` only
when it is valid (when its an `Element` instance).