containingText selection constraint
This adds the ability to restrict your selection to the first element that contains the given text.

```js
assert.dom('button').containingText('submit').hasAttribute('disabled');
```

This gives a reasonable replacement for the never-standardized `:contains()` selector that was implemented [in jQuery](https://api.jquery.com/contains-selector/). I have been missing it.